move clone_replace to a separate file
