+++ pytensor/ifelse.py
@@ -20,8 +20,9 @@ import pytensor.tensor as at
 from pytensor import as_symbolic
 from pytensor.compile import optdb
 from pytensor.configdefaults import config
-from pytensor.graph.basic import Apply, Variable, clone_replace, is_in_ancestors
+from pytensor.graph.basic import Apply, Variable, is_in_ancestors
 from pytensor.graph.op import _NoPythonOp
+from pytensor.graph.replace import clone_replace
 from pytensor.graph.rewriting.basic import GraphRewriter, in2out, node_rewriter
 from pytensor.graph.type import HasDataType, HasShape
 from pytensor.tensor.shape import Reshape, Shape, SpecifyShape, Unbroadcast
