+++ tests/graph/test_basic.py
@@ -4,7 +4,7 @@ from itertools import count
 import numpy as np
 import pytest
 
-from pytensor import config, function, shared
+from pytensor import shared
 from pytensor import tensor as at
 from pytensor.graph.basic import (
     Apply,
@@ -15,7 +15,6 @@ from pytensor.graph.basic import (
     as_string,
     clone,
     clone_get_equiv,
-    clone_replace,
     equal_computations,
     general_toposort,
     get_var_by_name,
@@ -30,18 +29,9 @@ from pytensor.graph.basic import (
 from pytensor.graph.op import Op
 from pytensor.graph.type import Type
 from pytensor.tensor.math import max_and_argmax
-from pytensor.tensor.type import (
-    TensorType,
-    dvector,
-    fvector,
-    iscalars,
-    matrix,
-    scalars,
-    vector,
-)
+from pytensor.tensor.type import TensorType, iscalars, matrix, scalars, vector
 from pytensor.tensor.type_other import NoneConst
 from pytensor.tensor.var import TensorVariable
-from tests import unittest_tools as utt
 from tests.graph.utils import MyInnerGraphOp
 
 
@@ -555,131 +545,6 @@ def test_get_var_by_name():
 
     exp_res = igo.fgraph.outputs[0]
     assert res == exp_res
-
-
-class TestCloneReplace:
-    def test_cloning_no_replace_strict_copy_inputs(self):
-        # This has nothing to do with scan, but it refers to the clone
-        # function that scan uses internally and that pfunc uses now and
-        # that users might want to use
-        x = vector("x")
-        y = vector("y")
-        z = shared(0.25)
-
-        f1 = z * (x + y) ** 2 + 5
-        f2 = clone_replace(f1, replace=None, rebuild_strict=True, copy_inputs_over=True)
-        f2_inp = graph_inputs([f2])
-
-        assert z in f2_inp
-        assert x in f2_inp
-        assert y in f2_inp
-
-    def test_cloning_no_replace_strict_not_copy_inputs(self):
-        # This has nothing to do with scan, but it refers to the clone
-        # function that scan uses internally and that pfunc uses now and
-        # that users might want to use
-        x = vector("x")
-        y = vector("y")
-        z = shared(0.25)
-
-        f1 = z * (x + y) ** 2 + 5
-        f2 = clone_replace(
-            f1, replace=None, rebuild_strict=True, copy_inputs_over=False
-        )
-        f2_inp = graph_inputs([f2])
-
-        assert z not in f2_inp
-        assert x not in f2_inp
-        assert y not in f2_inp
-
-    def test_cloning_replace_strict_copy_inputs(self):
-        # This has nothing to do with scan, but it refers to the clone
-        # function that scan uses internally and that pfunc uses now and
-        # that users might want to use
-        x = vector("x")
-        y = vector("y")
-        y2 = vector("y2")
-        z = shared(0.25)
-
-        f1 = z * (x + y) ** 2 + 5
-        f2 = clone_replace(
-            f1, replace={y: y2}, rebuild_strict=True, copy_inputs_over=True
-        )
-        f2_inp = graph_inputs([f2])
-        assert z in f2_inp
-        assert x in f2_inp
-        assert y2 in f2_inp
-
-    def test_cloning_replace_not_strict_copy_inputs(self):
-        # This has nothing to do with scan, but it refers to the clone
-        # function that scan uses internally and that pfunc uses now and
-        # that users might want to use
-        x = vector("x")
-        y = fvector("y")
-        y2 = dvector("y2")
-        z = shared(0.25)
-
-        f1 = z * (x + y) ** 2 + 5
-        f2 = clone_replace(
-            f1, replace={y: y2}, rebuild_strict=False, copy_inputs_over=True
-        )
-        f2_inp = graph_inputs([f2])
-        assert z in f2_inp
-        assert x in f2_inp
-        assert y2 in f2_inp
-
-    def test_cloning_replace_strict_not_copy_inputs(self):
-        # This has nothing to do with scan, but it refers to the clone
-        # function that scan uses internally and that pfunc uses now and
-        # that users might want to use
-        x = vector("x")
-        y = vector("y")
-        y2 = vector("y2")
-        z = shared(0.25)
-
-        f1 = z * (x + y) ** 2 + 5
-        f2 = clone_replace(
-            f1, replace=[(y, y2)], rebuild_strict=True, copy_inputs_over=False
-        )
-        f2_inp = graph_inputs([f2])
-        assert z not in f2_inp
-        assert x not in f2_inp
-        assert y2 not in f2_inp
-
-    def test_cloning_replace_not_strict_not_copy_inputs(self):
-        # This has nothing to do with scan, but it refers to the clone
-        # function that scan uses internally and that pfunc uses now and
-        # that users might want to use
-        x = vector("x")
-        y = fvector("y")
-        y2 = dvector("y2")
-        z = shared(0.25)
-
-        f1 = z * (x + y) ** 2 + 5
-        f2 = clone_replace(
-            f1, replace=[(y, y2)], rebuild_strict=False, copy_inputs_over=False
-        )
-        f2_inp = graph_inputs([f2])
-        assert z not in f2_inp
-        assert x not in f2_inp
-        assert y2 not in f2_inp
-
-    def test_clone(self):
-        def test(x, y, mention_y):
-            if mention_y:
-                d = 0.1 + 0 * y
-            else:
-                d = 0.1
-            out = clone_replace(y, replace={x: x + d})
-            return function([], out)()
-
-        x = shared(np.asarray(0.0, dtype=config.floatX))
-        utt.assert_allclose(
-            test(x, at.sum((x + 1) ** 2), mention_y=False), 1.21000003815
-        )
-        utt.assert_allclose(
-            test(x, at.sum((x + 1) ** 2), mention_y=True), 1.21000003815
-        )
 
 
 def test_clone_new_inputs():
