+++ tests/graph/test_replace.py
@@ -1,0 +1,135 @@
+import numpy as np
+import pytest
+
+import pytensor.tensor as pt
+from pytensor import config, function, shared
+from pytensor.graph.basic import graph_inputs
+from pytensor.graph.replace import clone_replace
+from pytensor.tensor import dvector, fvector, vector
+from tests import unittest_tools as utt
+from tests.graph.utils import MyOp, MyVariable
+
+
+class TestCloneReplace:
+    def test_cloning_no_replace_strict_copy_inputs(self):
+        # This has nothing to do with scan, but it refers to the clone
+        # function that scan uses internally and that pfunc uses now and
+        # that users might want to use
+        x = vector("x")
+        y = vector("y")
+        z = shared(0.25)
+
+        f1 = z * (x + y) ** 2 + 5
+        f2 = clone_replace(f1, replace=None, rebuild_strict=True, copy_inputs_over=True)
+        f2_inp = graph_inputs([f2])
+
+        assert z in f2_inp
+        assert x in f2_inp
+        assert y in f2_inp
+
+    def test_cloning_no_replace_strict_not_copy_inputs(self):
+        # This has nothing to do with scan, but it refers to the clone
+        # function that scan uses internally and that pfunc uses now and
+        # that users might want to use
+        x = vector("x")
+        y = vector("y")
+        z = shared(0.25)
+
+        f1 = z * (x + y) ** 2 + 5
+        f2 = clone_replace(
+            f1, replace=None, rebuild_strict=True, copy_inputs_over=False
+        )
+        f2_inp = graph_inputs([f2])
+
+        assert z not in f2_inp
+        assert x not in f2_inp
+        assert y not in f2_inp
+
+    def test_cloning_replace_strict_copy_inputs(self):
+        # This has nothing to do with scan, but it refers to the clone
+        # function that scan uses internally and that pfunc uses now and
+        # that users might want to use
+        x = vector("x")
+        y = vector("y")
+        y2 = vector("y2")
+        z = shared(0.25)
+
+        f1 = z * (x + y) ** 2 + 5
+        f2 = clone_replace(
+            f1, replace={y: y2}, rebuild_strict=True, copy_inputs_over=True
+        )
+        f2_inp = graph_inputs([f2])
+        assert z in f2_inp
+        assert x in f2_inp
+        assert y2 in f2_inp
+
+    def test_cloning_replace_not_strict_copy_inputs(self):
+        # This has nothing to do with scan, but it refers to the clone
+        # function that scan uses internally and that pfunc uses now and
+        # that users might want to use
+        x = vector("x")
+        y = fvector("y")
+        y2 = dvector("y2")
+        z = shared(0.25)
+
+        f1 = z * (x + y) ** 2 + 5
+        f2 = clone_replace(
+            f1, replace={y: y2}, rebuild_strict=False, copy_inputs_over=True
+        )
+        f2_inp = graph_inputs([f2])
+        assert z in f2_inp
+        assert x in f2_inp
+        assert y2 in f2_inp
+
+    def test_cloning_replace_strict_not_copy_inputs(self):
+        # This has nothing to do with scan, but it refers to the clone
+        # function that scan uses internally and that pfunc uses now and
+        # that users might want to use
+        x = vector("x")
+        y = vector("y")
+        y2 = vector("y2")
+        z = shared(0.25)
+
+        f1 = z * (x + y) ** 2 + 5
+        f2 = clone_replace(
+            f1, replace=[(y, y2)], rebuild_strict=True, copy_inputs_over=False
+        )
+        f2_inp = graph_inputs([f2])
+        assert z not in f2_inp
+        assert x not in f2_inp
+        assert y2 not in f2_inp
+
+    def test_cloning_replace_not_strict_not_copy_inputs(self):
+        # This has nothing to do with scan, but it refers to the clone
+        # function that scan uses internally and that pfunc uses now and
+        # that users might want to use
+        x = vector("x")
+        y = fvector("y")
+        y2 = dvector("y2")
+        z = shared(0.25)
+
+        f1 = z * (x + y) ** 2 + 5
+        f2 = clone_replace(
+            f1, replace=[(y, y2)], rebuild_strict=False, copy_inputs_over=False
+        )
+        f2_inp = graph_inputs([f2])
+        assert z not in f2_inp
+        assert x not in f2_inp
+        assert y2 not in f2_inp
+
+    def test_clone(self):
+        def test(x, y, mention_y):
+            if mention_y:
+                d = 0.1 + 0 * y
+            else:
+                d = 0.1
+            out = clone_replace(y, replace={x: x + d})
+            return function([], out)()
+
+        x = shared(np.asarray(0.0, dtype=config.floatX))
+        utt.assert_allclose(
+            test(x, pt.sum((x + 1) ** 2), mention_y=False), 1.21000003815
+        )
+        utt.assert_allclose(
+            test(x, pt.sum((x + 1) ** 2), mention_y=True), 1.21000003815
+        )
